// frontend/src/app/core/api/model/teamCreateDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TeamCreateDto { 
    name: string;
    description: string;
    department: TeamCreateDtoDepartmentEnum;
    leaderId?: number;
    expiresAt?: string;
    memberIds?: Array<number>;
}
export enum TeamCreateDtoDepartmentEnum {
    It = 'IT',
    Rd = 'RD',
    Hr = 'HR',
    Sales = 'SALES',
    Legal = 'LEGAL'
};